fix: refactor simulation, mint source module, cherry-picked
