+++ x/farming/simulation/operations.go
@@ -10,11 +10,11 @@ import (
 	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
 	"github.com/cosmos/cosmos-sdk/x/simulation"
 	utils "github.com/cosmosquad-labs/squad/types"
+	minttypes "github.com/cosmosquad-labs/squad/x/mint/types"
 
 	appparams "github.com/cosmosquad-labs/squad/app/params"
 	farmingkeeper "github.com/cosmosquad-labs/squad/x/farming/keeper"
 	farmingtypes "github.com/cosmosquad-labs/squad/x/farming/types"
-	liquiditytypes "github.com/cosmosquad-labs/squad/x/liquidity/types"
 )
 
 // Simulation operation weights constants.
@@ -33,6 +33,20 @@ var (
 			Denom:  "stake",
 			Amount: sdk.NewInt(1000),
 		},
+	}
+)
+
+var (
+	poolCoinDenoms = []string{
+		"pool93E069B333B5ECEBFE24C6E1437E814003248E0DD7FF8B9F82119F4587449BA5",
+		"pool3036F43CB8131A1A63D2B3D3B11E9CF6FA2A2B6FEC17D5AD283C25C939614A8C",
+		"poolE4D2617BFE03E1146F6BBA1D9893F2B3D77BA29E7ED532BB721A39FF1ECC1B07",
+	}
+
+	testCoinDenoms = []string{
+		"testa",
+		"testb",
+		"testc",
 	}
 )
 
@@ -121,7 +135,7 @@ func SimulateMsgCreateFixedAmountPlan(ak farmingtypes.AccountKeeper, bk farmingt
 		name := "simulation-test-" + simtypes.RandStringOfLength(r, 5) // name must be unique
 		creatorAcc := account.GetAddress()
 		// mint pool coins to simulate the real-world cases
-		funds, err := fundBalances(ctx, r, bk, creatorAcc)
+		funds, err := fundBalances(ctx, r, bk, creatorAcc, testCoinDenoms)
 		if err != nil {
 			return simtypes.NoOpMsg(farmingtypes.ModuleName, farmingtypes.TypeMsgCreateFixedAmountPlan, "unable to mint pool coins"), nil, nil
 		}
@@ -180,7 +194,7 @@ func SimulateMsgCreateRatioPlan(ak farmingtypes.AccountKeeper, bk farmingtypes.B
 		name := "simulation-test-" + simtypes.RandStringOfLength(r, 5) // name must be unique
 		creatorAcc := account.GetAddress()
 		// mint pool coins to simulate the real-world cases
-		_, err := fundBalances(ctx, r, bk, creatorAcc)
+		_, err := fundBalances(ctx, r, bk, account.GetAddress(), testCoinDenoms)
 		if err != nil {
 			return simtypes.NoOpMsg(farmingtypes.ModuleName, farmingtypes.TypeMsgCreateRatioPlan, "unable to mint pool coins"), nil, nil
 		}
@@ -377,41 +391,18 @@ func SimulateMsgHarvest(ak farmingtypes.AccountKeeper, bk farmingtypes.BankKeepe
 	}
 }
 
-// mintPoolCoins mints random amount of coins with the provided pool coin denoms and
+// fundBalances mints random amount of coins with the provided coin denoms and
 // send them to the simulated account.
-func mintPoolCoins(ctx sdk.Context, r *rand.Rand, bk farmingtypes.BankKeeper, acc simtypes.Account) (mintCoins sdk.Coins, err error) {
-	for _, denom := range []string{
-		"pool93E069B333B5ECEBFE24C6E1437E814003248E0DD7FF8B9F82119F4587449BA5",
-		"pool3036F43CB8131A1A63D2B3D3B11E9CF6FA2A2B6FEC17D5AD283C25C939614A8C",
-		"poolE4D2617BFE03E1146F6BBA1D9893F2B3D77BA29E7ED532BB721A39FF1ECC1B07",
-	} {
+func fundBalances(ctx sdk.Context, r *rand.Rand, bk farmingtypes.BankKeeper, acc sdk.AccAddress, denoms []string) (mintCoins sdk.Coins, err error) {
+	for _, denom := range denoms {
 		mintCoins = mintCoins.Add(sdk.NewInt64Coin(denom, int64(simtypes.RandIntBetween(r, 1e14, 1e15))))
 	}
 
-	if err := bk.MintCoins(ctx, liquiditytypes.ModuleName, mintCoins); err != nil {
+	if err := bk.MintCoins(ctx, minttypes.ModuleName, mintCoins); err != nil {
 		return nil, err
 	}
 
-	if err := bk.SendCoinsFromModuleToAccount(ctx, liquiditytypes.ModuleName, acc.Address, mintCoins); err != nil {
-		return nil, err
-	}
-	return mintCoins, nil
-}
-
-func fundBalances(ctx sdk.Context, r *rand.Rand, bk farmingtypes.BankKeeper, acc sdk.AccAddress) (mintCoins sdk.Coins, err error) {
-	for _, denom := range []string{
-		"testa",
-		"testb",
-		"testc",
-	} {
-		mintCoins = mintCoins.Add(sdk.NewInt64Coin(denom, int64(simtypes.RandIntBetween(r, 1e14, 1e17))))
-	}
-
-	if err := bk.MintCoins(ctx, liquiditytypes.ModuleName, mintCoins); err != nil {
-		return nil, err
-	}
-
-	if err := bk.SendCoinsFromModuleToAccount(ctx, liquiditytypes.ModuleName, acc, mintCoins); err != nil {
+	if err := bk.SendCoinsFromModuleToAccount(ctx, minttypes.ModuleName, acc, mintCoins); err != nil {
 		return nil, err
 	}
 	return mintCoins, nil
