+++ x/farming/simulation/proposals.go
@@ -7,7 +7,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
 	"github.com/cosmos/cosmos-sdk/x/simulation"
-	liquiditytypes "github.com/cosmosquad-labs/squad/x/liquidity/types"
+	minttypes "github.com/cosmosquad-labs/squad/x/mint/types"
 
 	"github.com/cosmosquad-labs/squad/app/params"
 	"github.com/cosmosquad-labs/squad/x/farming/keeper"
@@ -62,7 +62,7 @@ func SimulateAddPublicPlanProposal(ak types.AccountKeeper, bk types.BankKeeper, 
 			return nil
 		}
 
-		poolCoins, err := mintPoolCoins(ctx, r, bk, simAccount)
+		poolCoins, err := fundBalances(ctx, r, bk, simAccount.Address, poolCoinDenoms)
 		if err != nil {
 			return nil
 		}
@@ -93,7 +93,7 @@ func SimulateModifyPublicPlanProposal(ak types.AccountKeeper, bk types.BankKeepe
 			return nil
 		}
 
-		poolCoins, err := mintPoolCoins(ctx, r, bk, simAccount)
+		poolCoins, err := fundBalances(ctx, r, bk, simAccount.Address, poolCoinDenoms)
 		if err != nil {
 			return nil
 		}
@@ -197,10 +197,10 @@ func SimulateAdvanceEpoch(k keeper.Keeper, bk types.BankKeeper) simtypes.Content
 			plan := plans[rand.Intn(plansLen)]
 			if plan.GetType() == types.PlanTypePrivate {
 				mintCoins := sdk.NewCoins(sdk.NewInt64Coin("farmingreward", int64(simtypes.RandIntBetween(r, 0, 1e15))))
-				if err := bk.MintCoins(ctx, liquiditytypes.ModuleName, mintCoins); err != nil {
+				if err := bk.MintCoins(ctx, minttypes.ModuleName, mintCoins); err != nil {
 					return nil
 				}
-				if err := bk.SendCoinsFromModuleToAccount(ctx, liquiditytypes.ModuleName, plan.GetFarmingPoolAddress(), mintCoins); err != nil {
+				if err := bk.SendCoinsFromModuleToAccount(ctx, minttypes.ModuleName, plan.GetFarmingPoolAddress(), mintCoins); err != nil {
 					return nil
 				}
 				fmt.Println("[minted farming pool]", mintCoins, plan.String())
